move OffsetStore to internal
* no need for user to use it directly
* rename slick.OffsetStore to SlickOffsetStore
